+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 ### Added
 - Add `jose_jws:supported_crits/0` function.
 
+### Fixed
+- Fix media type without type returns an error.
+
 ## [0.1.0] (2021-01-18)
 ### Added
 - Add JWA HMAC with SHA2 algorithms